Released v2.0.
- Shuffled all of the core functionality to a static utility class.
- Made Laravel an optional dependency.
- Made ext-gmp an optional dependency.
- Reincluded support for PHP 7.0 to allow support for Laravel 5.5 LTS.
