+++ src/UuidModelAuthModel.php
@@ -1,0 +1,30 @@
+<?php declare(strict_types=1);
+
+/**
+ * This file is part of ConciseUUID, a PHP Experts, Inc., Project.
+ *
+ * Copyright © 2018-2019 PHP Experts, Inc.
+ * Author: Theodore R. Smith <theodore@phpexperts.pro>
+ *  GPG Fingerprint: 4BF8 2613 1C34 87AC D28F  2AD8 EB24 A91D D612 5690
+ *  https://www.phpexperts.pro/
+ *  https://github.com/phpexpertsinc/Zuora-API-Client
+ *
+ * This file is licensed under the MIT License.
+ */
+
+namespace PHPExperts\ConciseUuid;
+
+use Illuminate\Auth\Authenticatable;
+use Illuminate\Auth\Passwords\CanResetPassword;
+use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
+use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
+use Illuminate\Contracts\Auth\CanResetPassword as CanResetPasswordContract;
+use Illuminate\Foundation\Auth\Access\Authorizable;
+
+abstract class UuidModelAuthModel extends ConciseUuidModel implements
+    AuthenticatableContract,
+    AuthorizableContract,
+    CanResetPasswordContract
+{
+    use Authenticatable, Authorizable, CanResetPassword;
+}
